requirements.txt
streamlit
qrcode[pil]
Pillow
# testing commit again
